+++ minijinja-contrib/src/lib.rs
@@ -36,6 +36,10 @@ pub fn add_to_environment(env: &mut Environment) {
     env.add_filter("filesizeformat", filters::filesizeformat);
     env.add_filter("truncate", filters::truncate);
     env.add_filter("wordcount", filters::wordcount);
+    #[cfg(feature = "wordwrap")]
+    {
+        env.add_filter("wordwrap", filters::wordwrap);
+    }
     #[cfg(feature = "datetime")]
     {
         env.add_filter("datetimeformat", filters::datetimeformat);